refactor: 全项目行为不变重构，并修复三处死代码与资源管理缺陷
行为不变：不新增功能、不改渲染输出、不改规划结果、不碰公开接口签名。

守住的硬约束（一字未动）：
- 三个计数器：多图门控与版式轮转按镜头序号 index，运镜轮转按
  single_image_cursor
- _quad_spread 与 _camera_quad 共用同一个 amount
- fontsdir 只接受一个目录
- 运镜进度表达式 clip((on-1)/(frames-1),0,1)

等价重构（59 处）：
- I001 / UP037：import 排序；去掉注解引号（涉及文件均带
  from __future__ import annotations，去引号是 no-op）
- F401 / F541：删除未使用的导入（json、SimpleNamespace、_MIN_ZOOM）、
  去掉没有占位符的 f 前缀
- UP035：Callable / Iterator 改从 collections.abc 导入
- RUF007：zip(x, x[1:]) → itertools.pairwise（严格等价）
- C405 / C416 / RUF046：集合字面量、dict()、去掉多余的 int()
- SIM117：合并嵌套 with
- BLE001：给两处「换下一个 provider」「导演失败回退规则导演」的有意兜底
  补说明性 noqa，与同文件兄弟处理器保持一致

缺陷修复（3 处，均无可观察行为差异）：
- models/llama_server.py：清理条件 if log is None and sink is not DEVNULL
  的两个子句互斥（log 为 None 时 sink 必为 DEVNULL），close() 是死代码。
  改为 if log is not None；并给 except OSError 路径补一次 close()。
  注：CPython 会在生成器帧销毁时 GC 回收并 flush，故无输出差异——这是
  潜在缺陷修复，不是崩溃级问题。
- renderer.py：删除死变量 frames（赋值后从未被读取，image 分支直接
  return，video 分支全程不用）
- models/transcriber.py：每轮迭代新建闭包 → 提为模块级 _field()，并保留
  dict→.get（缺失返 None）/ 对象→getattr（缺失抛 AttributeError）的语义差异

验证：uv run pytest -q → 378 passed；uv run ruff check → 0 errors

+++ beatforge/models/vision_index.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-import hashlib
 import gc
+import hashlib
 import math
 import subprocess
 import sys
@@ -13,7 +13,6 @@ from tqdm.auto import tqdm
 
 from beatforge.media import MediaAsset, estimate_focus_point
 from beatforge.runtime import command
-
 
 # The Qwen-VL family of processors resizes every image to its own budget before the
 # vision tower sees anything - ``max_pixels = 1280 * 28 * 28`` is the usual default,
@@ -36,7 +35,7 @@ _CACHE_QUALITY = 95
 
 def _even(value: float) -> int:
     """Nearest even integer, at least 2. yuv420p and most patch grids want even sizes."""
-    return max(2, int(round(value / 2)) * 2)
+    return max(2, round(value / 2) * 2)
 
 
 def _fit_within(width: int, height: int, max_pixels: int) -> tuple[int, int]:
